Create ES-module project structure
