+++ js/data.js
@@ -1,0 +1,50 @@
+const TOTAL_POSTS_COUNT = 25;
+const MIN_LIKES_COUNT = 10;
+const MAX_LIKES_COUNT = 200;
+const TOTAL_COMMENTS_COUNT = 40;
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Пол',
+  'Дэвид',
+  'Боб',
+  'Дженис',
+  'Фрэнк',
+  'Йен',
+];
+
+const SURNAMES = [
+  'Маккартни',
+  'Гилмор',
+  'Марли',
+  'Джоплин',
+  'Заппа',
+  'Андерсен',
+];
+
+const DESCRIPTIONS = [
+  'black subbath',
+  'red wave',
+  'blue system',
+  'yellow river',
+  'green sleeves',
+];
+
+export {
+  TOTAL_POSTS_COUNT,
+  MIN_LIKES_COUNT,
+  MAX_LIKES_COUNT,
+  TOTAL_COMMENTS_COUNT,
+  COMMENTS,
+  NAMES,
+  SURNAMES,
+  DESCRIPTIONS
+};
